+++ .php-cs-fixer.php
@@ -8,10 +8,12 @@ $finder = PhpCsFixer\Finder::create()
 return (new PhpCsFixer\Config())
     ->setRules([
         '@PER-CS2.0' => true,
+        'binary_operator_spaces' => ['default' => 'single_space'],
         'blank_line_after_opening_tag' => true,
         'declare_strict_types' => true,
         'fully_qualified_strict_types' => ['import_symbols' => true],
         'global_namespace_import' => ['import_classes' => true, 'import_constants' => false, 'import_functions' => false],
+        'no_extra_blank_lines' => ['tokens' => ['attribute', 'case', 'continue', 'curly_brace_block', 'default', 'extra', 'parenthesis_brace_block', 'square_brace_block', 'switch']],
         'no_unused_imports' => true,
         'ordered_imports' => ['sort_algorithm' => 'alpha'],
         'phpdoc_line_span' => ['property' => 'single', 'method' => 'single'],